% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cnv_ICGC}
\alias{cnv_ICGC}
\title{Copy number variation profile of ICGC-OV.}
\format{
A data frame with copy number variation values, samples in columns, genes in rows, rownames corresponding to gene symbols.
}
\source{
\url{https://dcc.icgc.org/projects/OV-AU}
}
\usage{
cnv_ICGC
}
\description{
A dataset containing example validation copy number variation data for testing.
}
\keyword{datasets}
