% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{methy_M_ICGC}
\alias{methy_M_ICGC}
\title{DNA methylation profile of ICGC-OV.}
\format{
A data frame with z-score scaled M values, samples in columns, genes in rows, rownames corresponding to gene symbols.
}
\source{
\url{https://dcc.icgc.org/projects/OV-AU}
}
\usage{
methy_M_ICGC
}
\description{
A dataset containing example validation DNA methylation data for testing.
}
\keyword{datasets}
