% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mir_ICGC}
\alias{mir_ICGC}
\title{microRNA expression profile of ICGC-OV.}
\format{
A data frame with z-score scaled values, samples in columns, genes in rows, rownames corresponding to mature microRNA IDs.
}
\source{
\url{https://dcc.icgc.org/projects/OV-AU}
}
\usage{
mir_ICGC
}
\description{
A dataset containing example validation microRNA expression data for testing.
}
\keyword{datasets}
