% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mut_ICGC}
\alias{mut_ICGC}
\title{Gene mutation profile of ICGC-OV.}
\format{
A data frame with mutation values, samples in columns, genes in rows, rownames corresponding to gene symbols.
}
\source{
\url{https://dcc.icgc.org/projects/OV-AU}
}
\usage{
mut_ICGC
}
\description{
A dataset containing example validation gene muation data for testing.
}
\keyword{datasets}
