% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rna_log_tpm_ICGC}
\alias{rna_log_tpm_ICGC}
\title{Gene expression profile of ICGC-OV.}
\format{
A data frame with z-score scaled values, samples in columns, genes in rows, rownames corresponding to gene ensembl IDs.
}
\source{
\url{https://dcc.icgc.org/projects/OV-AU}
}
\usage{
rna_log_tpm_ICGC
}
\description{
A dataset containing example validation gene expression data for testing.
}
\keyword{datasets}
