% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TCGA_projection_mx}
\alias{TCGA_projection_mx}
\title{Projection matrixs derived from 226 paired mRNA expression, microRNA expression, DNA methylation, copy number variation, and mutation data from TCGA-OV samples.}
\format{
A list derived from sparse mCCA using PMA::MultiCCA()
}
\usage{
TCGA_projection_mx
}
\description{
A list containing projection matrixs of each omics. In the list, under ws, the five elements correspond to the projection matrices for mRNA, microRNA, methylation, copy number variation, and mutation.
}
\keyword{datasets}
