% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TCGAcnvscaled}
\alias{TCGAcnvscaled}
\title{Copy number variation profile of TCGA-OV.}
\format{
A DNA methylation profile with 5000 gene symbols and 226 samples.
}
\usage{
TCGAcnvscaled
}
\description{
A dataset containing the pre-processed gene-level copy number variation profile of 226 TCGA high-grade serous ovarian cancer samples.
}
\keyword{datasets}
