% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TCGAmetscaled}
\alias{TCGAmetscaled}
\title{DNA methylation profile of TCGA-OV.}
\format{
A DNA methylation profile with 2000 gene symbols and 226 samples.
}
\usage{
TCGAmetscaled
}
\description{
A dataset containing the pre-processed DNA methylation profile of 226 TCGA high-grade serous ovarian cancer samples.
}
\keyword{datasets}
