% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TCGAmiRNAscaled}
\alias{TCGAmiRNAscaled}
\title{microRNA exoression profile of TCGA-OV.}
\format{
A microRNA exoression profile with 100 microRNAs and 226 samples.
}
\usage{
TCGAmiRNAscaled
}
\description{
A dataset containing the pre-processed microRNA exoression profile of 226 TCGA high-grade serous ovarian cancer samples.
}
\keyword{datasets}
