% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TCGAmRNAscaled}
\alias{TCGAmRNAscaled}
\title{Gene expression profile of TCGA-OV.}
\format{
A microRNA exoression profile with 5000 mRNAs and 226 samples.
}
\usage{
TCGAmRNAscaled
}
\description{
A dataset containing the pre-processed mRNA exoression profile of 226 TCGA high-grade serous ovarian cancer samples.
}
\keyword{datasets}
