% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TCGAmutscaled}
\alias{TCGAmutscaled}
\title{Mutation profile of TCGA-OV.}
\format{
A DNA methylation profile with 200 gene symbols and 226 samples.
}
\usage{
TCGAmutscaled
}
\description{
A dataset containing the pre-processed gene mutation profile of 226 TCGA high-grade serous ovarian cancer samples.
}
\keyword{datasets}
